Fix build warnings: sign-conversion in allpass init, hidden processBlock overload
Cosmetic cleanup only, no behavior change. HilbertShifter::makeChain indexed
a std::array with int (-Wsign-conversion); QuaternionAudioProcessor's float
processBlock override hid the base class's double overload
(-Woverloaded-virtual) since we only support float processing -- pulled the
base overloads into scope explicitly instead of silently shadowing them.

+++ Source/DSP/HilbertShifter.h
@@ -103,7 +103,7 @@ private:
     static std::array<AllpassZ2, 4> makeChain(const double* coefs)
     {
         std::array<AllpassZ2, 4> c;
-        for (int i = 0; i < 4; ++i) c[i].a2 = coefs[i] * coefs[i];
+        for (size_t i = 0; i < 4; ++i) c[i].a2 = coefs[i] * coefs[i];
         return c;
     }
